# api/openapi-spec-v1.yml
openapi: 3.1.0
info:
  title: CBR/CBZ Comic Viewer API
  version: 1.0.0
  description: |
    REST API for comic file management and text-to-speech processing.
    Supports authentication, file upload/retrieval, reading progress tracking,
    and asynchronous TTS generation with LLM-based OCR.
  contact:
    name: API Support
    email: support@cbrviewer.example.com
servers:
  - url: http://localhost:8080
    description: Development server
security:
  - sessionAuth: []
    csrfToken: []
...
paths:
...
  /api/ratings/{fileId}:
    post:
      tags:
        - Ratings
      summary: Submit or update rating
      description: Submit a new rating or update existing rating for a comic file (1-5 stars)
      operationId: submitRating
      parameters:
        - name: fileId
          in: path
          description: Comic file ID to rate
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RatingRequest'
            example:
              rating: 4
      responses:
        '200':
          description: Rating saved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RatingResponse'
        '400':
          description: Invalid rating value (must be 1-5)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: VALIDATION_ERROR
                message: Rating must be between 1 and 5
                timestamp: '2025-01-06T14:50:00Z'
        '401':
          description: User not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    get:
      tags:
        - Ratings
      summary: Get user's rating
      description: Retrieve current user's rating for a comic file
      operationId: getUserRating
      parameters:
        - name: fileId
          in: path
          description: Comic file ID
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: User's rating retrieved (null if not rated)
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/RatingResponse'
                  - type: 'null'
        '401':
          description: User not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/ratings/{fileId}/average:
    get:
      tags:
        - Ratings
      summary: Get average rating
      description: Get aggregate rating statistics for a comic file
      operationId: getAverageRating
      security: []
      parameters:
        - name: fileId
          in: path
          description: Comic file ID
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Average rating retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AverageRatingResponse'
components:
  securitySchemes:
    sessionAuth:
      type: apiKey
      in: cookie
      name: JSESSIONID
      description: Spring Security session cookie (HttpOnly, Secure, SameSite=Lax)
    csrfToken:
      type: apiKey
      in: header
      name: X-CSRF-TOKEN
      description: CSRF protection token (required for state-changing operations)
  schemas:
...
    RatingRequest:
      type: object
      description: Submit or update rating request
      properties:
        rating:
          type: integer
          minimum: 1
          maximum: 5
          description: Rating value (1-5 stars)
          example: 4
      required:
        - rating
    RatingResponse:
      type: object
      description: Rating data
      properties:
        id:
          type: integer
          format: int64
          example: 123
        userId:
          type: integer
          format: int64
          example: 1
        fileId:
          type: integer
          format: int64
          example: 42
        rating:
          type: integer
          minimum: 1
          maximum: 5
          example: 4
        createdAt:
          type: string
          format: date-time
          example: '2025-01-06T15:00:00Z'
      required:
        - id
        - userId
        - fileId
        - rating
        - createdAt
    AverageRatingResponse:
      type: object
      description: Aggregate rating statistics for a comic file
      properties:
        fileId:
          type: integer
          format: int64
          example: 42
        averageRating:
          type: number
          format: double
          description: Average rating (1-5), null if no ratings
          example: 4.5
          nullable: true
        totalRatings:
          type: integer
          description: Total number of ratings
          example: 10
      required:
        - fileId
        - averageRating
        - totalRatings
    ErrorResponse:
      type: object
      description: Standard error response format
      properties:
        error:
          type: string
          description: Error code identifier
          example: AUTHENTICATION_FAILED
        message:
          type: string
          description: Human-readable error message
          example: Invalid username or password
        timestamp:
          type: string
          format: date-time
          example: '2025-01-06T14:32:00Z'
        path:
          type: string
          description: Request path that generated error
          example: /api/auth/login
      required:
        - error
        - message
        - timestamp